new update

+++ SOURCE/app/src/main/java/com/example/koichung/Util/Util.java
@@ -42,7 +42,7 @@ public class Util {
         return date;
     }
 
-    public static String formatMoney(Integer money) {
+    public static String formatMoney(long money) {
         DecimalFormat format = new DecimalFormat("###,###,###");
         return format.format(money);
     }
